requirements.txt
pandas==1.5.3
numpy==1.24.3
scipy==1.10.1
scikit-learn==1.2.2
scikit-image==0.20.0
albumentations==1.3.0
opencv-python==4.7.0.72
transformers==4.27.1
wandb==0.14.0
matplotlib==3.7.1
seaborn==0.12.2
timm==0.3.2
einops==0.3.0
h5py==3.1.0
jupyterlab==3.6.3
gdown==4.7.1
PyTurboJPEG==1.7.1
